Handle login page&operation in AccountController + redirect users after auth
