+++ CQuerMVC/Views/Account/Login.cshtml
@@ -1,20 +1,21 @@
-﻿@model Common.DTOs.LoginDto
+﻿@model LoginViewModel
 @{
     ViewData["Title"] = "Home Page";
 }
 <div id="login-row" class="row justify-content-center align-items-center">
     <div id="login-column" class="col-md-6">
         <div id="login-box" class="col-md-12">
-            <form asp-controller="Home" asp-action="Login" method="post" class="form">
+            <form asp-controller="Account" asp-action="Login" method="post" class="form">
+                <input type="hidden" asp-for="RedirectUrl">
                 <h1 class="text-center">Login</h1>
                 <p class="text-secondary"> Enter your login and password!</p>
                 <div class="form-group">
-                    <input type="text" class="form-control" asp-for="Username" placeholder="Username" required="required">
-                    <span asp-validation-for="Username" class="text-danger"></span>
+                    <input type="text" class="form-control" asp-for="LoginDto.Username" placeholder="Username" required="required">
+                    <span asp-validation-for="LoginDto.Username" class="text-danger"></span>
                 </div>
                 <div class="form-group">
-                    <input type="password" class="form-control" asp-for="Password" placeholder="Password" required="required">
-                    <span asp-validation-for="Password" class="text-danger"></span>
+                    <input type="password" class="form-control" asp-for="LoginDto.Password" placeholder="Password" required="required">
+                    <span asp-validation-for="LoginDto.Password" class="text-danger"></span>
                 </div>
                 <input type="submit" name="" value="Login" href="#">
             </form>
