+++ CQuerMVC/Views/Account/Register.cshtml
@@ -5,7 +5,7 @@
 <div id="login-row" class="row justify-content-center align-items-center">
     <div id="login-column" class="col-md-6">
         <div id="login-box" class="col-md-12">
-            <form asp-controller="Home" asp-action="Register" method="post" class="form">
+            <form asp-controller="Account" asp-action="Register" method="post" class="form">
                 <h1 class="text-center">Register</h1>
                 <p class="text-secondary"> Create new account</p>
                 <div class="form-group">
